added search endpoints using capability and talent id
