+++ contributions/api/controllers/contributions.py
@@ -16,7 +16,7 @@ import json
 import datetime
 import logging
 
-from flask import jsonify, request, g
+from flask import wrappers, request
 from bson import ObjectId
 
 import controllers.configs as cfg
@@ -25,6 +25,7 @@ import utils.datasetutils as datasetutils
 import utils.rest_handlers as rs_handlers
 import utils.mongoutils as mongoutils
 import utils.otherutils as otherutils
+import utils.otherutils as modelutils
 
 from utils import query_params
 from models.contribution import Contribution
@@ -84,7 +85,7 @@ def post(token_info):
             return rs_handlers.bad_request(msg_json)
 
         # check if the logged in user's login is included in contribution admin list
-        is_admin_user = check_login_admin(token_info["login"], contribution_admins)
+        is_admin_user = otherutils.check_login_admin(token_info["login"], contribution_admins)
         if not is_admin_user:
             msg = {
                 "reason": "Contribution admin list must contain logged in user",
@@ -99,7 +100,7 @@ def post(token_info):
         try:
             contributors_json = in_json["contributors"]
             for i in range(len(contributors_json)):
-                contributor, rest_contributor_json, msg = construct_contributors(contributors_json[i])
+                contributor, rest_contributor_json, msg = modelutils.construct_contributors(contributors_json[i])
                 if contributor is None:
                     return rs_handlers.bad_request(msg)
                 contributors_list.append(contributor)
@@ -112,7 +113,7 @@ def post(token_info):
         try:
             capability_json = in_json["capabilities"]
             for i in range(len(capability_json)):
-                capability, rest_capability_json, msg = construct_capability(capability_json[i])
+                capability, rest_capability_json, msg = modelutils.construct_capability(capability_json[i])
                 if capability is None:
                     return rs_handlers.bad_request(msg)
 
@@ -142,7 +143,7 @@ def post(token_info):
         try:
             talent_json = in_json["talents"]
             for i in range(len(talent_json)):
-                talent, rest_talent_json, msg = construct_talent(talent_json[i])
+                talent, rest_talent_json, msg = modelutils.construct_talent(talent_json[i])
                 if talent is None:
                     return rs_handlers.bad_request(msg)
 
@@ -167,7 +168,7 @@ def post(token_info):
                 if talent.requiredCapabilities is not None:
                     required_cap_list = talent.requiredCapabilities
                     for capability_json in required_cap_list:
-                        capability, rest_capability_json, msg = construct_capability(capability_json)
+                        capability, rest_capability_json, msg = modelutils.construct_capability(capability_json)
                         is_uuid = otherutils.check_if_uuid(capability.id)
                         if not is_uuid:
                             msg = {
@@ -201,7 +202,7 @@ def search(token_info=None, name=None):
     is_list = False
     query = dict()
 
-    login_id, is_login = get_login(token_info)
+    login_id, is_login = otherutils.get_login(token_info)
 
     # building query parematers
     query = query_params.format_query_status_login(query, login_id, is_login)
@@ -249,7 +250,7 @@ def search(token_info=None, name=None):
     return out_json
 
 def get(token_info=None, id=None):
-    login_id, is_login = get_login(token_info)
+    login_id, is_login = otherutils.get_login(token_info)
 
     data_list, is_objectid, is_error, resp = otherutils.get_data_list(id, login_id, is_login, coll_contribution)
 
@@ -290,7 +291,7 @@ def put(token_info, id):
     contribution_admins = contribution_dataset.contributionAdmins
 
     # check if the logged in user's login is included in contribution admin list
-    is_admin_user = check_login_admin(token_info["login"], contribution_admins)
+    is_admin_user = otherutils.check_login_admin(token_info["login"], contribution_admins)
     if not is_admin_user:
         msg = {
             "reason": "Contribution admin list must contain logged in user",
@@ -312,7 +313,7 @@ def put(token_info, id):
     try:
         capability_json = in_json["capabilities"]
         for i in range(len(capability_json)):
-            capability, rest_capability_json, msg = construct_capability(capability_json[i])
+            capability, rest_capability_json, msg = modelutils.construct_capability(capability_json[i])
             if capability is None:
                 return rs_handlers.bad_request(msg)
             # check if the capability id is in uuid format
@@ -335,7 +336,7 @@ def put(token_info, id):
     try:
         talent_json = in_json["talents"]
         for i in range(len(talent_json)):
-            talent, rest_talent_json, msg = construct_talent(talent_json[i])
+            talent, rest_talent_json, msg = modelutils.construct_talent(talent_json[i])
             if talent is None:
                 return rs_handlers.bad_request(msg)
             # check if the talent id is in uuid format
@@ -353,7 +354,7 @@ def put(token_info, id):
             if talent.requiredCapabilities is not None:
                 required_cap_list = talent.requiredCapabilities
                 for capability_json in required_cap_list:
-                    capability, rest_capability_json, msg = construct_capability(capability_json)
+                    capability, rest_capability_json, msg = modelutils.construct_capability(capability_json)
                     is_uuid = otherutils.check_if_uuid(capability.id)
                     if not is_uuid:
                         msg = {
@@ -388,7 +389,7 @@ def put(token_info, id):
     return out_json
 
 def delete(token_info, id):
-    login_id, is_login = get_login(token_info)
+    login_id, is_login = otherutils.get_login(token_info)
     data_list, is_objectid, is_error, resp = otherutils.get_data_list(id, login_id, is_login, coll_contribution)
 
     if is_error:
@@ -397,7 +398,7 @@ def delete(token_info, id):
     contribution_admins = data_list[0]['contributionAdmins']
 
     # check if the logged in user's login is included in contribution admin list
-    is_admin_user = check_login_admin(token_info["login"], contribution_admins)
+    is_admin_user = otherutils.check_login_admin(token_info["login"], contribution_admins)
     if not is_admin_user:
         msg = {
             "reason": "Contribution admin list must contain logged in user",
@@ -434,7 +435,7 @@ def allcapabilitiessearch(token_info=None, name=None):
     query = dict()
     is_list = False
 
-    login_id, is_login = get_login(token_info)
+    login_id, is_login = otherutils.get_login(token_info)
 
     query = query_params.format_query_status_login(query, login_id, is_login)
 
@@ -523,7 +524,7 @@ def allcapabilitiessearch(token_info=None, name=None):
     return return_json
 
 def capabilities_search(token_info=None, id=None):
-    login_id, is_login = get_login(token_info)
+    login_id, is_login = otherutils.get_login(token_info)
     contribution_dataset = None
 
     contribution_dataset, status_code = mongoutils.get_contribution_dataset_from_objectid(
@@ -587,7 +588,7 @@ def alltalentssearch(token_info=None, name=None):
     query = dict()
     is_list = False
 
-    login_id, is_login = get_login(token_info)
+    login_id, is_login = otherutils.get_login(token_info)
 
     query = query_params.format_query_status_login(query, login_id, is_login)
 
@@ -676,7 +677,7 @@ def alltalentssearch(token_info=None, name=None):
     return return_json
 
 def talents_search(token_info=None, id=None):
-    login_id, is_login = get_login(token_info)
+    login_id, is_login = otherutils.get_login(token_info)
 
     contribution_dataset, status_code = mongoutils.get_contribution_dataset_from_objectid(
         coll_contribution, id, login_id, is_login)
@@ -736,87 +737,80 @@ def talents_search(token_info=None, id=None):
 
     return talent_dataset
 
-def construct_capability(in_json):
-    # new installation of the app
-    capability_dataset = Capability('')
-    capability_dataset, restjson = datasetutils.update_capability_dataset_from_json(capability_dataset, in_json)
+def capabilities_get(token_info=None,id=None, capabilityid=None):
+    capability_datasets = capabilities_search(token_info, id)
 
-    return capability_dataset, restjson, None
+    # when the capablity_datasets is an error response
+    if isinstance(capability_datasets, wrappers.Response):
+        return capability_datasets
 
-def construct_talent(in_json):
-    is_required_field = True
-    error_required = ""
-    try:
-        error_required = "name"
-        name = in_json["name"]
-        error_required = "shortDescription"
-        description = in_json["shortDescription"]
-    except:
+    # otherwise it means that the capability_datasets contains the correct content.
+
+    # search capability that is the correct id
+    capability_index = 0
+    capability = None
+
+    for capability_dataset in capability_datasets:
+        if capability_dataset["id"] == capabilityid:
+            capability = capability_dataset
+            capability_index += 1
+
+    if capability_index > 1:
         msg = {
-            "reason": "Some of the required field in talent is not provided: " + str(error_required),
+            "reason": "There is more than one Capabilities with given capability id: "
+                      + str(capabilityid),
             "error": "Bad Request: " + request.url,
         }
-        msg_json = jsonutils.create_log_json("Contribution", "POST", msg)
-        logging.error("Contribution POST " + json.dumps(msg_json))
-        return None, None, msg_json
+        msg_json = jsonutils.create_log_json("Capability", "GET", msg)
+        logging.error("Capability GET " + json.dumps(msg_json))
+        return rs_handlers.bad_request(msg_json)
+    elif capability_index == 0:
+        msg = {
+            "reason": "There is no Capability with given capability id: "
+                      + str(capabilityid),
+            "error": "Not Found: " + request.url,
+        }
+        msg_json = jsonutils.create_log_json("Capability", "GET", msg)
+        logging.error("Capability GET " + json.dumps(msg_json))
+        return rs_handlers.not_found(msg_json)
+    else:
+        return capability
 
-    # new installation of the app
-    talent_dataset = Talent('')
-    talent_dataset, restjson = datasetutils.update_talent_dataset_from_json(talent_dataset, in_json)
+def talents_get(token_info=None,id=None, talentid=None):
+    talent_datasets = talents_search(token_info, id)
 
-    return talent_dataset, restjson, None
+    # when the capablity_datasets is an error response
+    if isinstance(talent_datasets, wrappers.Response):
+        return talent_datasets
 
-def construct_contributors(in_json):
-    # need to know if it is person or organization
-    # TODO make better algorithm for this, but for now use if there is firstname or lastname it is a Person, otherwise, organization
+    # otherwise it means that the talent_datasets contains the correct content.
 
-    is_person = False
-    try:
-        firstname = in_json["firstName"]
-        is_person = True
-    except:
-        pass
-    try:
-        lastname = in_json["lastName"]
-        is_person = True
-    except:
-        pass
+    # search talent that is the correct id
+    talent_index = 0
+    talent = None
 
-    contributor_dataset = None
-    if is_person:
-        contributor_dataset = Person('')
-        contributor_dataset, restjson = datasetutils.update_person_dataset_from_json(contributor_dataset, in_json)
-        # set affilication
-        try:
-            affilication_json = in_json['affiliation']
-            affilication_dataset = Organization('')
-            affilication_dataset, restjson = datasetutils.update_organization_dataset_from_json(affilication_dataset, affilication_json)
-            contributor_dataset.set_affilication(affilication_dataset)
-            del restjson['affiliation']
-        except:
-            pass
-    else: # organization
-        contributor_dataset = Organization('')
-        contributor_dataset, restjson = datasetutils.update_organization_dataset_from_json(contributor_dataset, in_json)
+    for talent_dataset in talent_datasets:
+        if talent_dataset["id"] == talentid:
+            talent = talent_dataset
+            talent_index += 1
 
-    return contributor_dataset, restjson, None
-
-def check_login_admin(login, inlist):
-    is_admin = False
-    for user in inlist:
-        if login == user:
-            is_admin = True
-
-    return is_admin
-
-def get_login(token_info):
-    is_login = False
-    login_id = ""
-
-    try:
-        login_id = token_info["login"]
-        is_login = True
-    except:
-        pass
-
-    return login_id, is_login
+    if talent_index > 1:
+        msg = {
+            "reason": "There is more than one Capabilities with given talent id: "
+                      + str(talentid),
+            "error": "Bad Request: " + request.url,
+        }
+        msg_json = jsonutils.create_log_json("Talent", "GET", msg)
+        logging.error("Talent GET " + json.dumps(msg_json))
+        return rs_handlers.bad_request(msg_json)
+    elif talent_index == 0:
+        msg = {
+            "reason": "There is no Capability with given capability id: "
+                      + str(talentid),
+            "error": "Not Found: " + request.url,
+        }
+        msg_json = jsonutils.create_log_json("Capability", "GET", msg)
+        logging.error("Capability GET " + json.dumps(msg_json))
+        return rs_handlers.not_found(msg_json)
+    else:
+        return talent
